% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cxxFlags.R
\name{cxxFlags}
\alias{cxxFlags}
\title{CXXFLAGS to compile with covafill}
\usage{
cxxFlags()
}
\value{
Returns a string with the CXXFLAGS needed to compile C++ code using covafill.
}
\description{
Get CXXFLAGS to compile with covafill
}
\examples{
\dontrun{
if(require("TMB")){
   f <- system.file("examples","tmbtest","tmbtest.cpp", package='covafillr')
   TMB::compile(f,CXXFLAGS = cxxFlags())
}
}
}
\author{
Christoffer Moesgaard Albertsen
}
\seealso{
\code{\link[TMB]{compile}}
}

