% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadJAGSModule.R
\name{loadJAGSModule}
\alias{loadJAGSModule}
\title{load JAGS module.}
\usage{
loadJAGSModule()
}
\value{
Nothing
}
\description{
Calls rjags::load.module with appropriate arguments to load the covafillr module.
}
\examples{
if(require("rjags") & covafillr:::.installed_with_jags)
   loadJAGSModule()
}
\seealso{
\code{\link[rjags]{load.module}}
}
\author{
Christoffer Moesgaard Albertsen
}
