% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{suggestBandwith}
\alias{suggestBandwith}
\title{Suggest bandwith for local polynomial regression}
\usage{
suggestBandwith(X, p)
}
\arguments{
\item{X}{A numeric matrix or vector of data coordinates}

\item{p}{Polynomial degree to base the suggestion on}
}
\value{
a vector or scalar of suggested bandwiths
}
\description{
The bandwith is suggested coordinate wise to be
\deqn{0.9 \sqrt{5} \min\left(sd(x),\frac{IQR(x)}{1.349}\right) n ^{-0.2}} (p+1)
where p is the polynomial degree used and n is the number of coordinate points.
}
\author{
Christoffer Moesgaard Albertsen
}
